requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
openai==1.3.5
chromadb==0.4.18
jinja2==3.1.2
pydantic==2.4.2
cryptography==41.0.5
numpy<2.0.0

sentence-transformers>=2.2.2
google-generativeai>=0.1.0
# google-genai
